+++ templates/part.navigation.starredfeed.php
@@ -11,7 +11,8 @@
 	<div class="app-navigation-entry-utils">
         <ul>
             <li class="app-navigation-entry-utils-counter"
-                ng-show="Navigation.getStarredCount() > 0">
+                ng-show="Navigation.getStarredCount() > 0"
+                title="{{ Navigation.getStarredCount() }}">
                 {{ Navigation.getStarredCount() | unreadCountFormatter }}
             </li>
         </ul>