+++ templates/part.navigation.unreadfeed.php
@@ -15,7 +15,8 @@
 	<div class="app-navigation-entry-utils">
         <ul>
             <li class="app-navigation-entry-utils-counter"
-                ng-show="Navigation.getUnreadCount() > 0">
+                ng-show="Navigation.getUnreadCount() > 0"
+                title="{{ Navigation.getUnreadCount() }}">
                 {{ Navigation.getUnreadCount() | unreadCountFormatter }}
             </li>
             <li class="app-navigation-entry-utils-menu-button"><button ng-click="optionsId = (optionsId == 'all' ? -1 : 'all')"></button></li>